src/requirements.txt
fastapi
uvicorn
pytest
httpx
flake8
celery
redis
black
ansible-runner[ansible]
sqlalchemy
psycopg2-binary
openai
anthropic
python-dotenv
jinja2
pyyaml
click
requests